Replacing 'jQuery' variable name with '$' sign
Currently in the ColVis plugin, in order to get the browser msie and
version, there were calls to jQuery.browser.msie and
jQuery.browser.version using the jQuery global namespace. While this
might not be an issue while loading the libraries externally, but if
your project has a RequireJS AMD setup then you would led to an error
saying jQuery is not defined. In order to fix this, we could use a $
sign instead of the jQuery namespace to get the browser msie and
version.